Remove public FlexContainer type
This only serves internal abstractions which are soon changing in a way that's incompatible with it.
