+++ redwood-layout-composeui/src/androidUnitTest/kotlin/app/cash/redwood/layout/composeui/ComposeUiFlexContainerTest.kt
@@ -24,7 +24,6 @@ import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import app.cash.redwood.layout.AbstractFlexContainerTest
 import app.cash.redwood.layout.TestFlexContainer
-import app.cash.redwood.layout.widget.FlexContainer
 import app.cash.redwood.layout.widget.Spacer
 import app.cash.redwood.snapshot.testing.ComposeSnapshotter
 import app.cash.redwood.snapshot.testing.ComposeUiTestWidgetFactory
@@ -72,7 +71,7 @@ class ComposeUiFlexContainerTest(
   class ComposeTestFlexContainer private constructor(
     private val delegate: ComposeUiFlexContainer,
   ) : TestFlexContainer<@Composable () -> Unit>,
-    FlexContainer<@Composable () -> Unit> by delegate {
+    YogaFlexContainer<@Composable () -> Unit> by delegate {
     private var childCount = 0
 
     override val children: ComposeWidgetChildren = delegate.children
